+++ test/0.8.25/vaults/contracts/StakingVault__HarnessForTestUpgrade.sol
@@ -108,7 +108,7 @@ contract StakingVault__HarnessForTestUpgrade is IStakingVault, OwnableUpgradeabl
 
     function rebalance(uint256 _ether) external {}
 
-    function report(uint256 _timestamp, uint256 _valuation, int256 _inOutDelta, uint256 _locked) external {}
+    function report(uint64 _timestamp, uint256 _valuation, int256 _inOutDelta, uint256 _locked) external {}
 
     function lock(uint256 _locked) external {}
